[Silver II] Title: 나무 자르기, Time: 1784 ms, Memory: 321268 KB -BaekjoonHub

+++ 백준/Silver/2805. 나무 자르기/나무 자르기.java
@@ -4,7 +4,7 @@ public class Main {
 	static int N, M;
 	static int[] length;
 	static int max = 0;
-	static long sol = 0;
+	static int sol = 0;
 
 	static void search(int start, int end) {
 		int mid = (start + end) / 2;
